refactoring to presenter
